repoview fixed

+++ plugin/app/plugins/pipeline_plugin/profiles/package/redhat/createrepo.sh
@@ -5,6 +5,7 @@ function createrepo(){
     eval "$REMOTE_COMMAND mkdir -p $directory" && \
       eval "$REMOTE_COMMAND /usr/bin/createrepo -s sha -d --update $directory"
     [[ $? != 0 ]] && return 1
-    which repoview 2>/dev/null && \
+    if which repoview 2>/dev/null; then
       eval "$REMOTE_COMMAND 'rm -rf $directory/repoview;/usr/bin/repoview $directory'"
+    fi
 }
